requirements.txt
streamlit==1.32.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
faker==20.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9
datetime
json5==0.9.14
pyyaml==6.0.1
tabulate==0.9.0
rich==13.7.0 